add searchResults page
